requirements.txt
django
djangorestframework
tablib==1.0.0
django-import-export==2.0.2
django-filter
requests
--no-binary=psycopg2
psycopg2
django-multiselectfield